Remove 'standard' from formatter list
It's confusing to mention it since 'prettier' now takes precedence
over it even for formatting JavaScript code. The code for running
'standard' is still present but not immediately usable without
knowledge of the internals of format-all. We need to think more about
what to do when there are multiple formatters for one language.