Fix compatibility with old line-number-at-pos function (#26)
Emacs 26 introduced a second parameter to the line-number-at-pos
function to ignore narrowing. We relied on that parameter, which meant
that format-all would signal an error in Emacs 24 and Emacs 25, even
though our Package-Requires tag states that those Emacs versions are
supported.

Fix by manually widening the buffer before saving the old line number.