Allow format-all-mode in buffers that don't have a formatter
Previouly format-all-mode would run use format-all-buffer as the
function to run in before-save-hook. Since format-all-buffer signals
an error when no formatter can be found for the buffer, this would
prevent the user from saving the buffer without first manually
disabling format-all-mode.

Fix by refactoring format-all-buffer into a separate interactive
command and hook function. The interactive command stays as before.
The hook function is lenient and simply returns nil instead of causing
an error when there is no formatter to be used.

Fixes issue #48. Is an alternative to PR #48.