Add --stderr to StandardRB formatter command (#157)
Without this, Rubocop (which is what StandardRB uses under the hood)
outputs everything to stdout, including a line of several "="
characters. StandardRB is smart enough to filter a majority of this
output, but not that line of "=" characters, which then makes its way to
the formatted file in emacs' buffer. This fixe the issue.

Closes issue #156.