Fix formatter-not-found error message
When we switched from `error` to `define-error` and `signal`, we
accidentally dropped the message about which formatter executable is
missing and how the user might be able to install it. Bring it back.

Unfortunately the formatting of the message is rather ugly, as Emacs
displays the information about the missing formatter in double quotes
after the constant part of the error message. I can't figure out how
to improve it.